feat(@dpc-sdp/ripple-tide-api): move epochToDate so build doesn't fail
