+++ packages/ripple-ui-core/src/index.ts
@@ -3,7 +3,7 @@ export { useAccessibleContainer } from './composables/useAccessibleContainer'
 export { formatDateRange } from './lib/formatDateRange'
 export { default as getGrantStatus } from './utils/getGrantStatus'
 export { isExternalUrl } from './utils/isExternalUrl'
-export { formatDate, epochToDate } from './lib/helpers'
+export { formatDate } from './lib/helpers'
 export { RplIconGroups } from './components/icon/constants'
 export { useRippleEvent } from './composables/useRippleEvent'
 export type { rplEventPayload } from './composables/useRippleEvent'
